+++ tests/testbed/011-table-name-override-through-demand-loading/post.rb
@@ -1,0 +1,9 @@
+class Post
+  self.table_name = "legacy_posts"
+end
+class Job
+  def run
+    p = Post.new
+    p.headline
+  end
+end